5315: add column ordering header menu items
