+++ beakerx/js/src/tableDisplay/tableScope.js
@@ -1509,239 +1509,8 @@ define([
       return items;
     };
 
-    var menuHelper = {
-      doAlignment: function(el, key) {
-        var container = el.closest('.bko-header-menu');
-        var colIdx = container.data('columnIndex');
-
-        //table variables
-        var table = $('#' + self.id).DataTable();
-        var bodyColumn = table.column(colIdx).nodes().to$();
-        var headerColumn = $(table.column(colIdx).header());
-        //remove align class
-        bodyColumn.removeClass('dtleft').removeClass('dtcenter').removeClass('dtright');
-        headerColumn.removeClass('dtleft').removeClass('dtcenter').removeClass('dtright');
-
-        //add align class
-        switch (key){
-          case 'L':
-            bodyColumn.addClass('dtleft');
-            headerColumn.addClass('dtleft');
-            break;
-          case 'C':
-            bodyColumn.addClass('dtcenter');
-            headerColumn.addClass('dtcenter');
-            break;
-          case 'R':
-            bodyColumn.addClass('dtright');
-            headerColumn.addClass('dtright');
-            break;
-        }
-
-        //update align
-        self.getCellAlign[self.colorder[colIdx] - 1] = key;
-        self.actualalign[self.colorder[colIdx] - 1] = key;
-        // bkSessionManager.setNotebookModelEdited(true); //TODO - check if needed
-      },
-      checkAlignment: function(container, key) {
-        var colIdx = container.data('columnIndex');
-        return self.actualalign[self.colorder[colIdx] - 1] === key;
-      },
-      doSorting: function(el, direction) {
-        var container = el.closest('.bko-header-menu');
-        var colIdx = container.data('columnIndex');
-
-        if (_.contains(['asc', 'desc'], direction)) {
-          self.table.order([colIdx, direction]).draw();
-        }
-      },
-      checkSorting: function(container, direction) {
-        var order = self.table.order();
-        var colIdx = container.data('columnIndex');
-
-        // server ordering
-        if (0 === order.length) {
-          return false;
-        }
-
-        if (_.contains(['asc', 'desc'], direction)) {
-          return (order[0][0] == colIdx && order[0][1] == direction);
-        } else {
-          return (order[0][0] !== colIdx);
-        }
-      },
-      doFixColumnLeft: function(el) {
-        var container = el.closest('.bko-header-menu');
-        var colIdx = container.data('columnIndex');
-        var fixed = this.isFixedLeft(container);
-        self.pagination.fixLeft = fixed ? 0 : colIdx;
-        self.applyChanges();
-      },
-      doFixColumnRight: function(el) {
-        var container = el.closest('.bko-header-menu');
-        var colIdx = container.data('columnIndex');
-        var fixed = this.isFixedRight(container);
-        self.pagination.fixRight = fixed ? 0 : self.columns.length - colIdx;
-        self.applyChanges();
-      },
-      isFixedRight: function(container) {
-        var colIdx = container.data('columnIndex');
-        return self.columns.length - colIdx === self.pagination.fixRight;
-      },
-      isFixedLeft: function(container) {
-        var colIdx = container.data('columnIndex');
-        return self.pagination.fixLeft === colIdx;
-      }
-    };
-
-    var headerMenuItems = {
-      items: [
-        {
-          title: 'Hide column',
-          action: function(el) {
-            var table = self.table;
-            var container = el.closest('.bko-header-menu');
-            var colIdx = container.data('columnIndex');
-            var column = table.column(colIdx);
-
-            column.visible(!column.visible());
-          }
-        },
-        {
-          title: 'Filter...',
-          icon: 'fa fa-filter',
-          tooltip: 'filter with an expression with a variable defined for each column and $ means the current column.  eg "$ > 5"',
-          action: function(el) {
-            var table = self.table;
-            var container = el.closest('.bko-header-menu');
-            var colIdx = container.data('columnIndex');
-            var column = table.column(colIdx);
-
-            self.doShowFilter(column, false);
-          }
-        },
-        {
-          title: 'Search...',
-          icon: 'fa fa-search',
-          tooltip: 'search this column for a substring',
-          action: function(el) {
-            var table = self.table;
-            var container = el.closest('.bko-header-menu');
-            var colIdx = container.data('columnIndex');
-            var column = table.column(colIdx);
-
-            self.doShowFilter(column, true);
-          }
-        },
-        {
-          title: 'Format',
-          action: null,
-          items: getFormatSubitems
-        },
-        {
-          title: 'Sort Ascending',
-          separator: true,
-          isChecked: function(container) {
-            return menuHelper.checkSorting(container, 'asc');
-          },
-          action: function(el) {
-            menuHelper.doSorting(el, 'asc');
-          }
-        },
-        {
-          title: 'Sort Descending',
-          isChecked: function(container) {
-            return menuHelper.checkSorting(container, 'desc');
-          },
-          action: function(el) {
-            menuHelper.doSorting(el, 'desc');
-          }
-        },
-        {
-          title: 'No Sort',
-          isChecked: function(container) {
-            return menuHelper.checkSorting(container);
-          },
-          action: function() {
-            self.table.order([0, 'asc']).draw();
-          }
-        },
-        {
-          title: 'Align Left',
-          separator: true,
-          isChecked: function(container) {
-            return menuHelper.checkAlignment(container, 'L');
-          },
-          action: function(el) {
-            menuHelper.doAlignment(el, 'L');
-          }
-        },
-        {
-          title: 'Align Center',
-          isChecked: function(container) {
-            return menuHelper.checkAlignment(container, 'C');
-          },
-          action: function(el) {
-            menuHelper.doAlignment(el, 'C');
-          }
-        },
-        {
-          title: 'Align Right',
-          isChecked: function(container) {
-            return menuHelper.checkAlignment(container, 'R');
-          },
-          action: function(el) {
-            menuHelper.doAlignment(el, 'R');
-          }
-        },
-        {
-          title: 'Heatmap',
-          shortcut: 'H',
-          separator: true,
-          isChecked: function(container) {
-            var colIdx = container.data('columnIndex');
-            var highlighter = self.cellHighlighters[self.colorder[colIdx]];
-            return highlighter && highlighter instanceof cellHighlighters.HeatmapHighlighter;
-          },
-          action: function(el) {
-            var container = el.closest('.bko-header-menu');
-            var colIdx = container.data('columnIndex');
-            self.showHideHeatmap(self.colorder[colIdx]);
-          }
-        },
-        {
-          title: 'Data Bars',
-          shortcut: 'B',
-          isChecked: function(container) {
-            var colIdx = container.data('columnIndex');
-            return self.barsOnColumn[self.colorder[colIdx]] === true;
-          },
-          action: function(el) {
-            var container = el.closest('.bko-header-menu');
-            var colIdx = container.data('columnIndex');
-            self.showHideBars(self.colorder[colIdx]);
-          }
-        },
-        {
-          title: 'Fix Left',
-          isChecked: function(container) {
-            return menuHelper.isFixedLeft(container);
-          },
-          action: function(el) {
-            menuHelper.doFixColumnLeft(el);
-          }
-        },
-        {
-          title: 'Fix Right',
-          isChecked: function(container) {
-            return menuHelper.isFixedRight(container);
-          },
-          action: function(el) {
-            menuHelper.doFixColumnRight(el);
-          }
-        }
-      ]
-    };
+    var createHeaderMenuItems = require('./tableHeaderMenu/createHeaderMenuItems');
+    var headerMenuItems = createHeaderMenuItems.call(self, cellHighlighters, getFormatSubitems);
 
     // build configuration
     var converter = self.allConverters[1];
